move add_dummy_state_and_pending_sign_requests helper to common.rs
